requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-filter==2.4.0
django-heroku==0.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
validate-docbr==1.8.2
whitenoise==5.2.0
